+++ todolist/src/main/java/study/todolist/repository/todo/TodoRepository.java
@@ -1,4 +1,4 @@
-package study.todolist.repository;
+package study.todolist.repository.todo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import study.todolist.entity.todo.TodoList;
